Cast return value to int
We're claiming to return an int, so we actually should do this.

+++ classes/Month.php
@@ -63,7 +63,7 @@ class Month
      */
     public function getLastDay()
     {
-        return date('t', $this->timestamp);
+        return (int) date('t', $this->timestamp);
     }
 
     /**
